// apps/airbyte/airbyte-protocol.gen.ts
export interface AirbyteStream {
  /**
   * Path to the field that will be used to determine if a record is new or modified since the
   * last sync. If not provided by the source, the end user will have to specify the
   * comparable themselves.
   */
  default_cursor_field?: string[]
  /**
   * Stream schema using Json Schema specs.
   */
  json_schema: {[key: string]: any}
  /**
   * Stream's name.
   */
  name: string
  /**
   * Optional Source-defined namespace. Currently only used by JDBC destinations to determine
   * what schema to write to. Airbyte streams from the same sources should have the same
   * namespace.
   */
  namespace?: string
  /**
   * If the source defines the cursor field, then any other cursor field inputs will be
   * ignored. If it does not, either the user_provided one is used, or the default one is used
   * as a backup.
   */
  source_defined_cursor?: boolean
  /**
   * If the source defines the primary key, paths to the fields that will be used as a primary
   * key. If not provided by the source, the end user will have to specify the primary key
   * themselves.
   */
  source_defined_primary_key?: Array<string[]>
  /**
   * List of sync modes supported by this stream.
   */
  supported_sync_modes: SyncMode[]
  [property: string]: any
}

export type SyncMode = 'full_refresh' | 'incremental'